+++ crates/miden-testing/src/kernel_tests/tx/test_account_interface.rs
@@ -1,9 +1,11 @@
+use alloc::vec::Vec;
+
 use assert_matches::assert_matches;
 use miden_lib::note::{create_p2id_note, create_p2ide_note};
 use miden_lib::transaction::TransactionKernel;
 use miden_objects::account::{Account, AccountId};
 use miden_objects::asset::FungibleAsset;
-use miden_objects::note::NoteType;
+use miden_objects::note::{Note, NoteType};
 use miden_objects::testing::account_id::{
     ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_IMMUTABLE_CODE,
     ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_UPDATABLE_CODE,
@@ -13,8 +15,9 @@ use miden_objects::testing::note::NoteBuilder;
 use miden_objects::{Felt, FieldElement, Word};
 use miden_tx::auth::UnreachableAuth;
 use miden_tx::{
-    NoteAccountExecution,
+    FailedNote,
     NoteConsumptionChecker,
+    NoteConsumptionInfo,
     TransactionExecutor,
     TransactionExecutorError,
 };
@@ -48,8 +51,9 @@ fn check_note_consumability_well_known_notes_success() -> anyhow::Result<()> {
         &mut RpoRandomCoin::new(Word::from([2u32; 4])),
     )?;
 
+    let notes = vec![p2id_note, p2ide_note];
     let tx_context = TransactionContextBuilder::with_existing_mock_account()
-        .extend_input_notes(vec![p2id_note, p2ide_note])
+        .extend_input_notes(notes.clone())
         .build()?;
     let source_manager = tx_context.source_manager();
 
@@ -69,13 +73,22 @@ fn check_note_consumability_well_known_notes_success() -> anyhow::Result<()> {
         tx_args,
         source_manager,
     )?;
-    assert_matches!(execution_check_result, NoteAccountExecution::Success);
 
+    assert_matches!(execution_check_result, NoteConsumptionInfo { successful, failed, .. }=> {
+    assert_eq!(successful.len(), notes.len());
+    successful.iter().zip(notes.iter()).for_each(|(success, note)| {
+        assert_eq!(success, note);
+    });
+    assert!(failed.is_empty());
+    });
     Ok(())
 }
 
+#[rstest::rstest]
+#[case::empty(vec![])]
+#[case::one(vec![create_p2any_note(ACCOUNT_ID_SENDER.try_into().unwrap(), &[FungibleAsset::mock(100)])])]
 #[test]
-fn check_note_consumability_custom_notes_success() -> anyhow::Result<()> {
+fn check_note_consumability_custom_notes_success(#[case] notes: Vec<Note>) -> anyhow::Result<()> {
     let tx_context = {
         let account = Account::mock(
             ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_UPDATABLE_CODE,
@@ -83,10 +96,8 @@ fn check_note_consumability_custom_notes_success() -> anyhow::Result<()> {
             Auth::IncrNonce,
             TransactionKernel::testing_assembler(),
         );
-        let input_note =
-            create_p2any_note(ACCOUNT_ID_SENDER.try_into().unwrap(), &[FungibleAsset::mock(100)]);
         TransactionContextBuilder::new(account)
-            .extend_input_notes(vec![input_note])
+            .extend_input_notes(notes.clone())
             .build()?
     };
     let source_manager = tx_context.source_manager();
@@ -107,8 +118,15 @@ fn check_note_consumability_custom_notes_success() -> anyhow::Result<()> {
         tx_args,
         source_manager,
     )?;
-    assert_matches!(execution_check_result, NoteAccountExecution::Success);
 
+    assert_matches!(execution_check_result, NoteConsumptionInfo { successful, failed, .. }=> {
+        if notes.is_empty() {
+            assert!(successful.is_empty());
+            assert!(failed.is_empty());
+        } else {
+            assert_eq!(successful.len(), notes.len());
+        }
+    });
     Ok(())
 }
 
@@ -181,18 +199,35 @@ fn check_note_consumability_failure() -> anyhow::Result<()> {
         input_notes,
         tx_args,
         source_manager,
-    )?;
+    );
 
-    assert_matches!(execution_check_result, NoteAccountExecution::Failure {
-        failed_note_id,
-        successful_notes,
-        error: Some(e)} => {
-            assert_eq!(failed_note_id, failing_note_2.id());
-            assert_eq!(successful_notes, [successful_note_2.id(),successful_note_1.id()].to_vec());
-            assert_matches!(e, TransactionExecutorError::TransactionProgramExecutionFailed(
-              ExecutionError::DivideByZero { .. }
-            ));
-        }
+    let execution_check_result = execution_check_result.unwrap();
+    assert_matches!(
+        execution_check_result,
+        NoteConsumptionInfo {
+            successful,
+            failed,
+            ..
+        } => {
+                assert_matches!(
+                    failed.first().expect("failed notes should exist"),
+                    FailedNote {
+                        note,
+                        error: TransactionExecutorError::TransactionProgramExecutionFailed(
+                            ExecutionError::DivideByZero { .. }),
+                        ..
+                    } => {
+                        assert_eq!(
+                            note.id(),
+                            failing_note_2.id(),
+                        );
+                    }
+                );
+                assert_eq!(
+                    [successful[0].id(), successful[1].id()],
+                    [successful_note_2.id(), successful_note_1.id()]
+                );
+            }
     );
     Ok(())
 }