+++ crates/miden-tx/src/executor/mod.rs
@@ -7,7 +7,7 @@ use miden_lib::transaction::TransactionKernel;
 use miden_objects::account::AccountId;
 use miden_objects::assembly::SourceManager;
 use miden_objects::block::{BlockHeader, BlockNumber};
-use miden_objects::note::{NoteId, NoteScript};
+use miden_objects::note::{Note, NoteScript};
 use miden_objects::transaction::{
     AccountInputs,
     ExecutedTransaction,
@@ -34,7 +34,38 @@ mod data_store;
 pub use data_store::DataStore;
 
 mod notes_checker;
-pub use notes_checker::{NoteConsumptionChecker, NoteInputsCheck};
+pub use notes_checker::NoteConsumptionChecker;
+
+// NOTE CONSUMPTION INFO
+// ================================================================================================
+
+/// Represents a failed note consumption.
+#[derive(Debug)]
+#[non_exhaustive]
+pub struct FailedNote {
+    pub note: Note,
+    pub error: TransactionExecutorError,
+}
+
+/// Contains information about the successful and failed consumption of notes.
+#[derive(Default, Debug)]
+#[non_exhaustive]
+pub struct NoteConsumptionInfo {
+    pub successful: Vec<Note>,
+    pub failed: Vec<FailedNote>,
+}
+
+impl NoteConsumptionInfo {
+    /// Creates a new [`NoteConsumptionInfo`] instance with the given successful notes.
+    pub fn new_successful(successful: Vec<Note>) -> Self {
+        Self { successful, ..Default::default() }
+    }
+
+    /// Creates a new [`NoteConsumptionInfo`] instance with the given successful and failed notes.
+    pub fn new(successful: Vec<Note>, failed: Vec<FailedNote>) -> Self {
+        Self { successful, failed }
+    }
+}
 
 // TRANSACTION EXECUTOR
 // ================================================================================================
@@ -289,9 +320,9 @@ where
     // CHECK CONSUMABILITY
     // ============================================================================================
 
-    /// Executes the transaction with specified notes, returning the [NoteAccountExecution::Success]
-    /// if all notes has been consumed successfully and [NoteAccountExecution::Failure] if some note
-    /// returned an error.
+    /// Validates input notes, transaction inputs, and account inputs before executing the
+    /// transaction with specified notes. Keeps track and returns both successfully consumed notes
+    /// as well as notes that failed to be consumed.
     ///
     /// The `source_manager` is used to map potential errors back to their source code. To get the
     /// most value out of it, use the source manager from the
@@ -302,7 +333,7 @@ where
     ///
     /// # Errors:
     /// Returns an error if:
-    /// - If required data can not be fetched from the [DataStore].
+    /// - If required data can not be fetched from the [`DataStore`].
     /// - If the transaction host can not be created from the provided values.
     /// - If the execution of the provided program fails on the stage other than note execution.
     #[maybe_async]
@@ -313,34 +344,36 @@ where
         notes: InputNotes<InputNote>,
         tx_args: TransactionArgs,
         source_manager: Arc<dyn SourceManager>,
-    ) -> Result<NoteAccountExecution, TransactionExecutorError> {
+    ) -> Result<NoteConsumptionInfo, TransactionExecutorError> {
+        if notes.is_empty() {
+            return Ok(NoteConsumptionInfo::default());
+        }
+        // Validate input notes.
         let mut ref_blocks = validate_input_notes(&notes, block_ref)?;
         ref_blocks.insert(block_ref);
 
+        // Validate account inputs.
         let (account, seed, ref_block, mmr) =
             maybe_await!(self.data_store.get_transaction_inputs(account_id, ref_blocks))
                 .map_err(TransactionExecutorError::FetchTransactionInputsFailed)?;
-
         validate_account_inputs(&tx_args, &ref_block)?;
 
+        // Prepare transaction inputs.
         let tx_inputs = TransactionInputs::new(account, seed, ref_block, mmr, notes)
             .map_err(TransactionExecutorError::InvalidTransactionInputs)?;
-
         let (stack_inputs, advice_inputs) =
             TransactionKernel::prepare_inputs(&tx_inputs, &tx_args, None)
                 .map_err(TransactionExecutorError::ConflictingAdviceMapEntry)?;
 
+        // Prepare host for transaction execution.
         let input_notes = tx_inputs.input_notes();
-
         let scripts_mast_store = ScriptMastForestStore::new(
             tx_args.tx_script(),
             input_notes.iter().map(|n| n.note().script()),
         );
-
         let acct_procedure_index_map =
             AccountProcedureIndexMap::from_transaction_params(&tx_inputs, &tx_args, &advice_inputs)
                 .map_err(TransactionExecutorError::TransactionHostCreationFailed)?;
-
         let mut host = TransactionExecutorHost::new(
             &tx_inputs.account().into(),
             input_notes.clone(),
@@ -349,10 +382,9 @@ where
             acct_procedure_index_map,
             self.authenticator,
         );
-
         let advice_inputs = advice_inputs.into_advice_inputs();
 
-        // execute the transaction kernel
+        // Execute the transaction kernel.
         let result = vm_processor::execute(
             &TransactionKernel::main(),
             stack_inputs,
@@ -363,32 +395,58 @@ where
         )
         .map_err(TransactionExecutorError::TransactionProgramExecutionFailed);
 
+        let (_, _, _, _, input_notes) = tx_inputs.into_parts();
         match result {
-            Ok(_) => Ok(NoteAccountExecution::Success),
-            Err(tx_execution_error) => {
+            Ok(_) => {
+                // Return all the input notes as successful.
+                Ok(NoteConsumptionInfo::new_successful(
+                    input_notes.into_iter().map(|note| note.into_note()).collect::<Vec<_>>(),
+                ))
+            },
+            Err(error) => {
                 let notes = host.tx_progress().note_execution();
 
-                // empty notes vector means that we didn't process the notes, so an error
-                // occurred somewhere else
+                // Empty notes vector means that we didn't process the notes, so an error
+                // occurred.
                 if notes.is_empty() {
-                    return Err(tx_execution_error);
+                    return Err(error);
                 }
 
                 let ((last_note, last_note_interval), success_notes) = notes
                     .split_last()
                     .expect("notes vector should not be empty because we just checked");
 
-                // if the interval end of the last note is specified, then an error occurred after
-                // notes processing
+                // If the interval end of the last note is specified, then an error occurred after
+                // notes processing.
                 if last_note_interval.end().is_some() {
-                    return Err(tx_execution_error);
+                    return Err(error);
                 }
 
-                Ok(NoteAccountExecution::Failure {
-                    failed_note_id: *last_note,
-                    successful_notes: success_notes.iter().map(|(note, _)| *note).collect(),
-                    error: Some(tx_execution_error),
-                })
+                // Partition the input notes into successful and failed results.
+                let mut successful = Vec::with_capacity(success_notes.len());
+                let mut failed = Vec::with_capacity(1);
+                for (i, note) in input_notes.into_iter().enumerate() {
+                    if i < success_notes.len() {
+                        debug_assert_eq!(
+                            success_notes[i].0,
+                            note.id(),
+                            "notes should be processed in the same order as they appear in the input notes"
+                        );
+                        successful.push(note.into_note());
+                    } else {
+                        // This is the last (failed) note.
+                        debug_assert_eq!(
+                            *last_note,
+                            note.id(),
+                            "notes should be processed in the same order as they appear in the input notes"
+                        );
+                        failed.push(FailedNote { note: note.into_note(), error });
+                        break;
+                    }
+                }
+
+                // Return information about all the consumed notes.
+                Ok(NoteConsumptionInfo::new(successful, failed))
             },
         }
     }
@@ -526,22 +584,4 @@ fn map_execution_error(exec_err: ExecutionError) -> TransactionExecutorError {
         },
         _ => TransactionExecutorError::TransactionProgramExecutionFailed(exec_err),
     }
-}
-
-// HELPER ENUM
-// ================================================================================================
-
-/// Describes whether a transaction with a specified set of notes could be executed against target
-/// account.
-///
-/// [NoteAccountExecution::Failure] holds data for error handling: `failing_note_id` is an ID of a
-/// failing note and `successful_notes` is a vector of note IDs which were successfully executed.
-#[derive(Debug)]
-pub enum NoteAccountExecution {
-    Success,
-    Failure {
-        failed_note_id: NoteId,
-        successful_notes: Vec<NoteId>,
-        error: Option<TransactionExecutorError>,
-    },
 }