+++ crates/miden-tx/src/lib.rs
@@ -12,10 +12,10 @@ mod executor;
 pub use executor::{
     DataStore,
     ExecutionOptions,
+    FailedNote,
     MastForestStore,
-    NoteAccountExecution,
     NoteConsumptionChecker,
-    NoteInputsCheck,
+    NoteConsumptionInfo,
     TransactionExecutor,
     TransactionExecutorHost,
 };